+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetInstrumentationOptions.cs
@@ -17,7 +17,6 @@
 using System;
 using System.Diagnostics;
 using System.Web;
-using OpenTelemetry.Context.Propagation;
 
 namespace OpenTelemetry.Instrumentation.AspNet
 {
@@ -26,12 +25,6 @@ namespace OpenTelemetry.Instrumentation.AspNet
     /// </summary>
     public class AspNetInstrumentationOptions
     {
-        /// <summary>
-        /// Gets or sets <see cref="TextMapPropagator"/> for context propagation.
-        /// By default, <see cref="Propagators.DefaultTextMapPropagator" /> will be used.
-        /// </summary>
-        public TextMapPropagator Propagator { get; set; }
-
         /// <summary>
         /// Gets or sets a Filter function to filter instrumentation for requests on a per request basis.
         /// The Filter gets the HttpContext, and should return a boolean.